Create a separate file for the Device object
The Device object is currently defined in ecephys.py although icephys,
ecephys and ophys use it. This is unfortunate and will be changed in
this commit.

Fix #443.
